fix(enrollment): split prepare locks by purchase flow
